Feature : 카카오 로그인 (토큰 발급), Redis Cache / Session 분할 구축

+++ backend/member/src/main/java/com/trendflow/member/auth/dto/authentication/KakaoAccess.java
@@ -11,7 +11,7 @@ public class KakaoAccess {
     private String accessToken;
     private String refreshToken;
     private String tokenType;
-    private Long accessTokenExpire;
-    private Long refreshTokenExpire;
+    private Integer accessTokenExpire;
+    private Integer refreshTokenExpire;
     private List<String> scope;
 }
